+++ apps/mobile/package.json
@@ -9,7 +9,6 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "typecheck": "tsc --pretty",
     "lint": "eslint . --ext .js,.ts,.tsx"
   },
   "dependencies": {
@@ -54,6 +53,6 @@
     "@lenstube/lens": "*",
     "babel-plugin-module-resolver": "^5.0.0",
     "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
+    "typescript": "^4.9.4"
   }
 }
